Count total_leios_bytes differently (#438)
* Count total_leios_bytes including the transactions generated

This seems to be the only way to get a total sum of things that will be
stored for FullWithoutIbs and FullWithTxReferences variants.

Open point: this will cound transaction bytes double for the other
variants and worsen theire spatial efficiency while we would not store
txs double of course.

* Actually count transaction bytes on the RB event

This should now be more honest about the things we actually need to store.

* Only count transaction sizes in variants where it matters

---------

Co-authored-by: Simon Gellis <simongellis@gmail.com>